docs(nixos): fix UEFI/GPT parted commands
From parted's info page:

```
Command: mkpart [part-type name fs-type] start end

part-type is one of ‘primary’, ‘extended’ or ‘logical’, and may be
specified only with ‘msdos’ or ‘dvh’ partition tables. A name must be
specified for a ‘gpt’ partition table.
```

https://www.gnu.org/software/parted/manual/parted.html#mkpart

The previous version of the commands either didn't work, or set 2 of the
partitions to have the name "primary".

I tried to make it clear which part of the command was the name by using
quotes.

I am torn on whether to keep the "ESP" name in GPT, or to change it to
"boot" to match the filesystem on it. I decided to change it to match.

I also changed the swapon command to use the names assigned earlier in the
guide.